+++ repository/database.ts
@@ -1,6 +1,7 @@
 import knex from "knex";
 import config from "./knexfile";
-import { AN1_COM_FOND, ActeLegislatif, AN1_COM_FOND_NOMIN } from "./Acts";
+import { AN1_COM_FOND, AN1_COM_FOND_NOMIN } from "./Acts";
+import { Dossier, ActeLegislatif, Organe, Acteur } from "./types";
 
 const db = knex(config.development);
 
@@ -16,156 +17,10 @@ export async function listTables() {
   }
 }
 
-export interface DossierRow {
-  uid: string;
-  xsiType: string;
-  legislature: string;
-  senatChemin: string;
-  titre: string;
-  titreChemin: string;
-  theme: string | null;
-  codeProcedure: string;
-  libelleProcedure: string;
-  causeFusionDossier: string | null;
-  dossierAbsorbantRefUid: string | null;
-  organeRefUid: string | null;
-}
-
-export interface ActLegislatif {
-  uid: string;
-  codeActe: string;
-  nomCanonique: string;
-  libelleCourtActe: string;
-  xsiType: string;
-  dateActe: Date | null;
-  organeRefUid: string;
-  organeProvenanceRefUid: string | null;
-  famCodeStatutConclusion: string | null;
-  libelleStatutConclusion: string | null;
-  famCodeCasSaisine: string | null;
-  libelleCasSaisine: string | null;
-  anneeDecisionConseilConstitutionnel: string | null;
-  urlConclusionConseilConstitutionnel: string | null;
-  motifConseilConstitutionnel: string | null;
-  numDecisionConseilConstitutionnel: string | null;
-  auteurMotionRefUid: string | null;
-  famCodeTypeMotion: string | null;
-  libelleTypeMotion: string | null;
-  famCodeTypeMotionCensure: string | null;
-  libelleTypeMotionCensure: string | null;
-  famCodeDecision: string | null;
-  libelleDecision: string | null;
-  formuleDecision: string | null;
-  codeLoiRefUid: string | null;
-  dateJo: string | null;
-  numJo: string | null;
-  referenceNor: string | null;
-  typeJo: string | null;
-  urlLegifrance: string | null;
-  titreLoi: string | null;
-  urlEcheancierLoi: string | null;
-  dateFermetureContributionInternaute: string | null;
-  dateOuvertureContributionInternaute: string | null;
-  dateJoce: string | null;
-  refJoce: string | null;
-  titreTexteEuropeen: string | null;
-  typeTexteEuropeen: string | null;
-  odjRefUid: string | null;
-  reunionRefUid: string | null;
-  famCodeStatutAdoption: string | null;
-  libelleStatutAdoption: string | null;
-  texteAdopteRefUid: string | null;
-  texteAssocieRefUid: string;
-  famCodeStatutTypeDeclaration: string | null;
-  libelleStatutTypeDeclaration: string | null;
-  dossierRefUid: string;
-}
-
-export type Document = {
-  titrePrincipal: string;
-  titrePrincipalCourt: string;
-  uid: string;
-  classeCode: string;
-  classeLibelle: string;
-  texteDeLoi: boolean;
-  depotCode: string;
-  depotLibelle: string;
-  especeCode: string;
-  especeLibelle: string;
-  sousTypeCode: string;
-  sousTypeLibelle: string;
-  sousTypeLibelleEdition: string;
-  statutAdoption: null | string;
-  niveauCorrection: null | string;
-  typeCorrection: null | string;
-  dateCreation: Date;
-  dateDepot: Date;
-  datePublication: Date;
-  datePublicationWeb: Date;
-  denominationStructurelle: string;
-  amendable: null | string;
-  dian: null | string;
-  isbn: null | string;
-  nbPage: null | string;
-  prix: null | string;
-  legislature: string;
-  adoptionConforme: boolean;
-  formule: string;
-  numNotice: string;
-  provenance: string;
-  xsiType: string;
-  auteurUid: string;
-  documentParentRefUid: null;
-};
-
-export type Organe = {
-  uid: string;
-  codeType: string;
-  libelle: string;
-  libelleEdition: string;
-  libelleAbrege: string;
-  libelleAbrev: string;
-  organeParentRefUid: null | string;
-  regime: null | string;
-  legislature: null | string;
-  secretaire01: null | string;
-  secretaire02: null | string;
-  regimeJuridique: null | string;
-  siteInternet: null | string;
-  nombreReunionsAnnuelles: null | string;
-  positionPolitique: null | string;
-  preseance: null | string;
-  couleurAssociee: null | string;
-  dateDebut: null | Date;
-  dateAgrement: null | Date;
-  dateFin: null | Date;
-  xsiType: string;
-};
-
-export type Acteur = {
-  uid: string;
-  prenom: string;
-  nom: string;
-  civ: string;
-  dateNais: Date;
-  dateDeces: null;
-  villeNais: string;
-  depNais: string;
-  paysNais: string;
-  profession: string;
-  catSocPro: string;
-  famSocPro: string;
-  slug: string;
-  uriHatvp: null | string;
-  deputeActif: boolean;
-  deputeGroupeParlementaireUid: null | string;
-  mandatPrincipalUid: null | string;
-};
-
 export async function getDossiers(
   { legislature = 16 },
   limit = 10
-): Promise<DossierRow[]> {
+): Promise<Dossier[]> {
   try {
     const rows = await db
       .select("*")
@@ -184,7 +39,7 @@ export async function getDossier(
   id: string
 ): Promise<
   | {
-      dossier: DossierRow;
+      dossier: Dossier;
       /**
        * uid de la commission saisie sur le fond.
        * Données à ércupérer dans organes.
@@ -322,10 +177,7 @@ export async function getDossier(
   }
 }
 
-export async function getTable(
-  table: string,
-  limit = 10
-): Promise<DossierRow[]> {
+export async function getTable(table: string, limit = 10): Promise<Dossier[]> {
   try {
     const rows = await db
       .select("*")
